+++ src/components/Avatar/Avatar.jsx
@@ -4,16 +4,20 @@ import PropTypes from "prop-types";
 import cx from "classnames";
 import { BEMClass } from "helpers/bem-helper";
 import { backwardCompatibilityForProperties } from "helpers/backwardCompatibilityForProperties";
-import { getElementColor, elementColorsNames } from "utils/colors-vars-map";
+import { elementColorsNames, getElementColor } from "utils/colors-vars-map";
 import { AVATAR_SIZES, AVATAR_TYPES } from "./AvatarConstants";
 import { AvatarBadge } from "./AvatarBadge";
 import { AvatarContent } from "./AvatarContent";
+import Tooltip from "../Tooltip/Tooltip";
+import ClickableWrapper from "../Clickable/ClickableWrapper";
+import Dialog from "../Dialog/Dialog";
 import "./Avatar.scss";
 
 const AVATAR_CSS_BASE_CLASS = "monday-style-avatar";
 const bemHelper = BEMClass(AVATAR_CSS_BASE_CLASS);
 
 const Avatar = ({
+  id,
   type,
   className,
   textClassName,
@@ -21,8 +25,10 @@ const Avatar = ({
   src,
   icon,
   text,
-  role,
+  tooltipProps,
   ariaLabel,
+  withoutTooltip,
+  role,
   backgroundColor,
   square,
   disabled,
@@ -38,7 +44,8 @@ const Avatar = ({
   bottomRightBadgeProps,
   withoutBorder,
   customSize,
-  customBackgroundColor
+  customBackgroundColor,
+  onClick
 }) => {
   const overrideSquare = backwardCompatibilityForProperties([square, isSquare]);
   const overrideDisabled = backwardCompatibilityForProperties([disabled, isDisabled], false);
@@ -47,8 +54,18 @@ const Avatar = ({
     return src ? {} : { backgroundColor: getElementColor(backgroundColor) };
   }, [src, backgroundColor, customBackgroundColor]);
   const sizeStyle = useMemo(() => {
-    return customSize ? { height: `${customSize}px`, width: `${customSize}px` } : {};
+    return customSize ? { height: customSize, width: customSize } : {};
   }, [customSize]);
+
+  const overrideTooltipProps = useMemo(() => {
+    if (withoutTooltip) return undefined;
+
+    if (tooltipProps) {
+      return { content: ariaLabel, ...tooltipProps };
+    } else {
+      return { content: ariaLabel };
+    }
+  }, [ariaLabel, tooltipProps, withoutTooltip]);
 
   const badgesContainer = useMemo(() => {
     const badges = [];
@@ -97,34 +114,40 @@ const Avatar = ({
   }, [size, topLeftBadgeProps, topRightBadgeProps, bottomLeftBadgeProps, bottomRightBadgeProps]);
 
   return (
-    <div className={cx(AVATAR_CSS_BASE_CLASS, className)}>
-      <div
-        className={cx(
-          bemHelper({ element: "circle" }),
-          bemHelper({ element: "circle", state: type }),
-          bemHelper({ element: "circle", state: size }),
-          {
-            [bemHelper({ element: "circle", state: "is-disabled" })]: overrideDisabled,
-            [bemHelper({ element: "circle", state: "is-square" })]: overrideSquare,
-            [bemHelper({ element: "circle", state: "without-border" })]: withoutBorder
-          }
-        )}
-        aria-hidden={ariaHidden}
-        tabIndex={tabIndex}
-        style={{ ...backgroundColorStyle, ...sizeStyle }}
+    <div id={id} className={cx(AVATAR_CSS_BASE_CLASS, className, bemHelper({ state: size }))} style={sizeStyle}>
+      <ClickableWrapper
+        isClickable={!!onClick}
+        clickableProps={{ onClick: onClick, tabIndex: "-1", className: bemHelper({ element: "clickableWrapper" }) }}
       >
-        <AvatarContent
-          type={type}
-          size={size}
-          textClassName={textClassName}
-          src={src}
-          icon={icon}
-          text={text}
-          ariaLabel={ariaLabel}
-          role={role}
-        />
-      </div>
-      {badgesContainer}
+        <Tooltip
+          showTrigger={[Dialog.hideShowTriggers.FOCUS, Dialog.hideShowTriggers.MOUSE_ENTER]}
+          hideTrigger={[Dialog.hideShowTriggers.BLUR, Dialog.hideShowTriggers.MOUSE_LEAVE]}
+          {...overrideTooltipProps}
+        >
+          <div
+            className={cx(bemHelper({ element: "circle" }), bemHelper({ element: "circle", state: type }), {
+              [bemHelper({ element: "circle", state: "is-disabled" })]: overrideDisabled,
+              [bemHelper({ element: "circle", state: "is-square" })]: overrideSquare,
+              [bemHelper({ element: "circle", state: "without-border" })]: withoutBorder
+            })}
+            aria-hidden={ariaHidden}
+            tabIndex={tabIndex}
+            style={{ ...backgroundColorStyle }}
+          >
+            <AvatarContent
+              type={type}
+              size={size}
+              src={src}
+              icon={icon}
+              text={text}
+              ariaLabel={ariaLabel}
+              role={role}
+              textClassName={textClassName}
+            />
+          </div>
+          {badgesContainer}
+        </Tooltip>
+      </ClickableWrapper>
     </div>
   );
 };
@@ -135,8 +158,12 @@ Avatar.colors = elementColorsNames;
 Avatar.backgroundColors = elementColorsNames;
 
 Avatar.propTypes = {
+  id: PropTypes.string,
   src: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
   text: PropTypes.string,
+  tooltipProps: PropTypes.shape(Tooltip.propTypes),
+  ariaLabel: PropTypes.string,
+  withoutTooltip: PropTypes.bool,
   icon: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
   type: PropTypes.oneOf([Avatar.types.TEXT, Avatar.types.ICON, Avatar.types.IMG]),
   className: PropTypes.string,
@@ -144,7 +171,6 @@ Avatar.propTypes = {
   backgroundColor: PropTypes.oneOf(Object.values(Avatar.colors)),
   customBackgroundColor: PropTypes.string,
   role: PropTypes.string,
-  ariaLabel: PropTypes.string,
   size: PropTypes.oneOf([Avatar.sizes.LARGE, Avatar.sizes.MEDIUM, Avatar.sizes.SMALL]),
   customSize: PropTypes.number,
   tabIndex: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
@@ -155,20 +181,26 @@ Avatar.propTypes = {
   topRightBadgeProps: PropTypes.object,
   bottomLeftBadgeProps: PropTypes.object,
   bottomRightBadgeProps: PropTypes.object,
-  withoutBorder: PropTypes.bool
+  withoutBorder: PropTypes.bool,
+  /**
+   * (event) => void
+   */
+  onClick: PropTypes.func
 };
-
 Avatar.defaultProps = {
+  id: undefined,
   src: undefined,
   className: "",
   textClassName: "",
   icon: undefined,
   text: undefined,
+  tooltipProps: undefined,
+  ariaLabel: undefined,
+  withoutTooltip: false,
   type: AVATAR_TYPES.TEXT,
   backgroundColor: elementColorsNames.CHILI_BLUE,
   customBackgroundColor: null,
   role: undefined,
-  ariaLabel: undefined,
   size: AVATAR_SIZES.LARGE,
   customSize: null,
   tabIndex: 0,
@@ -179,7 +211,8 @@ Avatar.defaultProps = {
   topRightBadgeProps: undefined,
   bottomLeftBadgeProps: undefined,
   bottomRightBadgeProps: undefined,
-  withoutBorder: false
+  withoutBorder: false,
+  onClick: undefined
 };
 
 export default Avatar;