+++ src/storybook/index.js
@@ -1,3 +1,4 @@
 export * from "./components";
 export * from "./decorators/vertical-stories";
+export * from "./decorators/memory-stats";
 export * from "./functions/create-component-story";